Expert_Tasks/Expert_Task02-03/backend/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==2.3.2
flask-cors==5.0.1
flatbuffers==25.2.10
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==20.1.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
keras==3.9.2
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml-dtypes==0.4.1
namex==0.0.9
numpy==1.26.0
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
Pillow==9.5.0
protobuf==5.29.4
Pygments==2.19.1
python-dotenv==1.0.0
requests==2.32.3
rich==14.0.0
setuptools==80.3.1
six==1.17.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
tensorflow_intel==2.18.0
termcolor==3.1.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==2.3.6
wheel==0.45.1
wrapt==1.17.2
